resolved some warnings (boxing, overriding tag)

+++ omovCore/trunk/src/main/java/net/sourceforge/omov/core/tools/vlc/WebinterfaceVlcPlayer.java
@@ -174,7 +174,7 @@ abstract class WebinterfaceVlcPlayer implements IVlcPlayer {
 		Integer id = getPlaylist().getIdByUri(uri);
 		if(id != null) {
 			try {
-				playById(id);
+				playById(id.intValue());
 			} catch (IOException e) {
 				throw new BusinessException("Could not play file '"+file.getAbsolutePath()+"'!", e);
 			}
